refactor: main filesystem code
