+++ src/main/app/index.js
@@ -4,9 +4,9 @@ import { exec } from 'child_process'
 import dayjs from 'dayjs'
 import log from 'electron-log'
 import { app, BrowserWindow, clipboard, dialog, ipcMain, systemPreferences } from 'electron'
+import { isChildOfDirectory } from 'common/filesystem/paths'
 import { isLinux, isOsx } from '../config'
 import parseArgs from '../cli/parser'
-import { isChildOfDirectory } from '../filesystem'
 import { normalizeMarkdownPath } from '../filesystem/markdown'
 import { getMenuItemById } from '../menu'
 import { selectTheme } from '../menu/actions/theme'
