+++ src/main/filesystem/index.js
@@ -1,156 +1,6 @@
 import fs from 'fs-extra'
 import path from 'path'
-import { hasMarkdownExtension } from '../utils'
-import { IMAGE_EXTENSIONS } from '../config'
-
-/**
- * Test whether or not the given path exists.
- *
- * @param {string} p The path to the file or directory.
- * @returns {boolean}
- */
-export const exists = async p => {
-  // fs.exists is deprecated.
-  try {
-    await fs.access(p)
-    return true
-  } catch(_) {
-    return false
-  }
-}
-
-/**
- * Ensure that a directory exist.
- *
- * @param {string} dirPath The directory path.
- */
-export const ensureDirSync = dirPath => {
-  try {
-    fs.ensureDirSync(dirPath)
-  } catch (e) {
-    if (e.code !== 'EEXIST') {
-      throw e
-    }
-  }
-}
-
-/**
- * Returns true if the path is a directory with read access.
- *
- * @param {string} dirPath The directory path.
- */
-export const isDirectory = dirPath => {
-  try {
-    return fs.existsSync(dirPath) && fs.lstatSync(dirPath).isDirectory()
-  } catch (_) {
-    return false
-  }
-}
-
-/**
- * Returns true if the path is a file with read access.
- *
- * @param {string} filepath The file path.
- */
-export const isFile = filepath => {
-  try {
-    return fs.existsSync(filepath) && fs.lstatSync(filepath).isFile()
-  } catch (_) {
-    return false
-  }
-}
-
-/**
- * Returns true if the path is a symbolic link with read access.
- *
- * @param {string} filepath The link path.
- */
-export const isSymbolicLink = filepath => {
-  try {
-    return fs.existsSync(filepath) && fs.lstatSync(filepath).isSymbolicLink()
-  } catch (_) {
-    return false
-  }
-}
-
-/**
- * Returns true if the path is a markdown file.
- *
- * @param {string} filepath The path or link path.
- */
-export const isMarkdownFile = filepath => {
-  return isFile(filepath) && hasMarkdownExtension(filepath)
-}
-/**
- * Returns ture if the path is an image file.
- *
- * @param {string} filepath The path
- */
-export const isImageFile = filepath => {
-  const extname = path.extname(filepath)
-  return isFile(filepath) && IMAGE_EXTENSIONS.some(ext => {
-    const EXT_REG = new RegExp(ext, 'i')
-    return EXT_REG.test(extname)
-  })
-}
-
-/**
- * Returns true if the path is a markdown file or symbolic link to a markdown file.
- *
- * @param {string} filepath The path or link path.
- */
-export const isMarkdownFileOrLink = filepath => {
-  if (!isFile(filepath)) return false
-  if (hasMarkdownExtension(filepath)) {
-    return true
-  }
-
-  // Symbolic link to a markdown file
-  if (isSymbolicLink(filepath)) {
-    const targetPath = fs.readlinkSync(filepath)
-    return isFile(targetPath) && hasMarkdownExtension(targetPath)
-  }
-  return false
-}
-
-/**
- * Check if the both paths point to the same file.
- *
- * @param {string} pathA The first path.
- * @param {string} pathB The second path.
- * @param {boolean} [isNormalized] Are both paths already normalized.
- */
-export const isSamePathSync = (pathA, pathB, isNormalized = false) => {
-  if (!pathA || !pathB) return false
-  const a = isNormalized ? pathA : path.normalize(pathA)
-  const b = isNormalized ? pathB : path.normalize(pathB)
-  if (a.length !== b.length) {
-    return false
-  } else if (a === b) {
-    return true
-  } else if (a.toLowerCase() === b.toLowerCase()) {
-    try {
-      const fiA = fs.statSync(a)
-      const fiB = fs.statSync(b)
-      return fiA.ino === fiB.ino
-    } catch (_) {
-      // Ignore error
-    }
-  }
-  return false
-}
-
-/**
- * Check whether a file or directory is a child of the given directory.
- *
- * @param {string} dir The parent directory.
- * @param {string} child The file or directory path to check.
- */
-export const isChildOfDirectory = (dir, child) => {
-  if (!dir || !child) return false
-  const relative = path.relative(dir, child)
-  return relative && !relative.startsWith('..') && !path.isAbsolute(relative)
-}
+import { isDirectory, isFile, isSymbolicLink } from 'common/filesystem'
 
 /**
  * Normalize the path into an absolute path and resolves the link target if needed.
