+++ src/main/filesystem/watcher.js
@@ -2,8 +2,9 @@ import path from 'path'
 import fs from 'fs-extra'
 import log from 'electron-log'
 import chokidar from 'chokidar'
-import { getUniqueId, hasMarkdownExtension } from '../utils'
-import { exists } from '../filesystem'
+import { exists } from 'common/filesystem'
+import { hasMarkdownExtension } from 'common/filesystem/paths'
+import { getUniqueId } from '../utils'
 import { loadMarkdownFile } from '../filesystem/markdown'
 import { isLinux } from '../config'
