+++ src/main/menu/index.js
@@ -2,8 +2,8 @@ import fs from 'fs'
 import path from 'path'
 import { app, ipcMain, Menu } from 'electron'
 import log from 'electron-log'
+import { ensureDirSync, isDirectory, isFile } from 'common/filesystem'
 import { isLinux } from '../config'
-import { ensureDirSync, isDirectory, isFile } from '../filesystem'
 import { parseMenu } from '../keyboard/shortcutHandler'
 import configureMenu, { configSettingMenu } from '../menu/templates'
