+++ src/renderer/store/editor.js
@@ -1,9 +1,9 @@
 import { clipboard, ipcRenderer, shell } from 'electron'
 import path from 'path'
 import equal from 'deep-equal'
+import { isSamePathSync } from 'common/filesystem/paths'
 import bus from '../bus'
 import { hasKeys, getUniqueId } from '../util'
-import { isSamePathSync } from '../util/fileSystem'
 import listToTree from '../util/listToTree'
 import { createDocumentState, getOptionsFromState, getSingleFileState, getBlankFileState } from './help'
 import notice from '../services/notification'
