+++ src/renderer/util/fileSystem.js
@@ -2,7 +2,7 @@ import path from 'path'
 import fse from 'fs-extra'
 import dayjs from 'dayjs'
 import Octokit from '@octokit/rest'
-import { isImageFile } from '../../main/filesystem'
+import { isImageFile } from 'common/filesystem/paths'
 import { dataURItoBlob, getContentHash } from './index'
 import axios from 'axios'
 
@@ -20,33 +20,6 @@ export const paste = ({ src, dest, type }) => {
 
 export const rename = (src, dest) => {
   return fse.move(src, dest)
-}
-
-/**
- * Check if the both paths point to the same file.
- *
- * @param {string} pathA The first path.
- * @param {string} pathB The second path.
- * @param {boolean} [isNormalized] Are both paths already normalized.
- */
-export const isSamePathSync = (pathA, pathB, isNormalized = false) => {
-  if (!pathA || !pathB) return false
-  const a = isNormalized ? pathA : path.normalize(pathA)
-  const b = isNormalized ? pathB : path.normalize(pathB)
-  if (a.length !== b.length) {
-    return false
-  } else if (a === b) {
-    return true
-  } else if (a.toLowerCase() === b.toLowerCase()) {
-    try {
-      const fiA = fse.statSync(a)
-      const fiB = fse.statSync(b)
-      return fiA.ino === fiB.ino
-    } catch (_) {
-      // Ignore error
-    }
-  }
-  return false
 }
 
 export const moveImageToFolder = async (pathname, image, dir) => {
